- SRV DNS request is now saved in cache and reused.
I use the minimum ttl of all SRV records, minored in a week (as of RFC-1035,
too big a ttl should not be used. A week is the example in the RFC. I used this
for my implementation).

- Small code improvement in dns results.

As I am not sure of the result in absence of SRV records (FALSE as an error
or empty array?), I test emptiness of result instead as boolean.