=== Rate This Page === Plugin Name: Rate This Page Plugin Contributors: Agents Of Value Donate link: http://www.agentsofvalue.com/2011/09/wikipedia-style-rate-this-page-plugin/ Tags: plugin wordpress,wordpress blog,page rating,page plugin,page rate,feedback tool,article feedback tool,wordpress rating,custom wordpress plugin,wordpress reviews plugin Requires at least: 3.0 Tested up to: 3.2.1 Stable Tag: version 1.0
== Description ==
Rate This Page plug-in allow the visitors to rate your article posts and pages and show rating report system. Plug-in supports the use of cookies and saving of data using ajax.

Supported language: English

== About Plugin Support ==
Plug-in is provided for free. However, the support is never free. Currently available step by step tutorials on the plugin website.

== Installation ==
Minimal Requirements

* PHP: 5.x.x
* mySQL: 4.0, 4.1 or 5.x
* WordPress: 3.0 or newer

Recommended Requirements

* PHP: 5.2.x or newer
* mySQL: 5.x
* WordPress: 3.0 or newer

Please see the step by step with screenshot on this post:  http://www.agentsofvalue.com/?p=6351

1. Upload 'rate-this-page' folder to the '/wp-content/plugins/' directory.
2. Activate the plugin through the 'Plugins' menu in WordPress see screenshot-2.
3. Go to the admin section of the plugin at the left sidebar on wp-admin see "Rate This Page".
4. then select "All Post" if you want the plugin to appear in all of your blog articles please see the Configuration Tab
then the Rate This Page tool rating will be present at all article posts.
5. Select "By Category" if you want the plugin to appear in selected categories of your choice.
6. Enjoy! (^_^)

== Frequently Asked Questions ==

Question:
How did the Rate This Page plug-in calculate the Average Rating in the rating report tab?

Answer:
The plug-in Average Rating formula:
If there are (5) people voted for trustworthy and rated 5.0, then the average rating is 5.0 because, the article's trustworthy is weighted by the number of people's ratings.
Now, if there are (5) people voted for trustworthy that rated 5.0, and another one person voted 1.0 rating, then the average rating is 4.3.
In short:
Total number of rating's Trustworthy, Objective, Complete and Well-written (of certain article post) is divided by the number of people who rated the article = Average Rating.


== Screenshots ==

1. Screenshot of How to activate the plugin: /rate-this-page/screenshot-1.png
2. Screenshot of How to access the Rate This Page Admin Area: /rate-this-page/screenshot-2.png 
3. Screenshot of How to display the plug-in either by "All Post" or "By Category": /rate-this-page/screenshot-3.png
4. Screenshot of Display the plug-in by using the category name please click the checkbox: /rate-this-page/screenshot-4.png
5. Screenshot of How to Edit the Article: /rate-this-page/screenshot-5.png
6. Screenshot of Access the Rating Report Tab: /rate-this-page/screenshot-6.png


== Changelog ==

= 1.0 =
New release: stable in displaying the rating in blog article posts or wordpress either in All Post or By Category selection.
The plugin only works on article blog post (not page) of wordpress. It can be shown either on "All Posts" and "By Category" selection.

== Upgrade Notice ==

= 1.1 =
New: added the ability to display the rating in Wordpress Selected Pages and selected By Category blog posts.
New: added the uninstall and reinstall features of the plug-in database tables.