Option removal and validation.
The option to enter the domain name as one of the parameters to GA
have been removed. It's better to retrieve the site URL and use that.

Functionality for retrieving the domain name from the full site URL
have been implemented within the code template. The function will
remove protocol prefixes for http and https, aswell as the trailing
slash (if any).

Implemented the functionality for validating the value given as the
GA property ID. If the value has a invalid format, a error message
will appear and the value will be nullified.