Tiled Galleries: Responsive and flexible.
Resize tiled galleries for smaller screens or whenever the width of the content area changes. It would be have been nice to handle this entirely in PHP and with percentage widths/heights, but given the way that height percentages work (and the need to maintain specific margin sizes), a JavaScript? solution is our best bet right now.

Tested on rectangular, square, and circle galleries; works at all widths from < 100px wide to > 2000px.