Added class to [bgmp-list] container to allow targeting multiple lists on one page
Previously the element only had an ID because it was assumed there would only ever be one list on a page, but it is possible for there to be multiple lists. Ideally it should only have a class, but the ID was left to preserve backwards-compatability.

See http://wordpress.org/support/topic/110-view-on-map-issue?replies=5#post-3996402 for the issue that prompted this change.