Prefetch media items when setting up the shortcode.
In the past, users would add the shortcode to a page, then view the page and have to wait 30+ seconds for a refresh to pull in content. This loads the content when the shortcode is added to the page, so that it will be available immediately.

Note that it often will appear to not be working, due to Twitter's search API constraints. See the note in prefetch_media_items().