Use get_current_blog_id() instead of undefined blog_id variable
The blog_id variable was not defined, but the statement was working because is_user_member_of_blog() checks for a correct value and defaults to get_current_blog_id() if it isn't passed one. So, this commit doesn't change the behavior of the statement, but it does avoid the PHP notice that gets thrown when the undefined blog_id is referenced.