Fire an AJAX request to store the category instead of hijacking the tag field.
Currently we hijack the tag field in order to send the category from the client side to the server, since P2 doesn't have any hooks to allow us to do it in a more elegant way.

The next version of p2 will have a p2_new_post_submit_success that fires when a new post is created. We can hook onto that and then send our own AJAX request to assign the category to the new post.